Adds an option to skip the Configuration of Cassandra

+++ cassandra/bin/initialize-entrypoint.sh
@@ -1,16 +1,18 @@
 #!/bin/bash
 set -e
 
-echo "Starting Casandra to initialize it..."
-cassandra
-echo "Waiting for Cassandra to be ready..."
-sleep 20
-echo "Initialising the database..."
-cqlsh -f /app/data/database.cql
-echo "Killing Cassandra..."
-pkill -f cassandra
-sleep 20
-echo "Cassandra database initialised..."
+if [ "${CASSANDRA_INITIALIZE}" = true ] ; then
+    echo "Starting Casandra to initialize it..."
+	cassandra -R
+	echo "Waiting for Cassandra to be ready..."
+	sleep 20
+	echo "Initialising the database..."
+	cqlsh -f /app/data/benchflow.cql
+	echo "Killing Cassandra..."
+	pkill -f cassandra
+	sleep 20
+	echo "Cassandra database initialised..."
+fi
 
 # execute the base image entrypoint in the context of the current shell
 source /docker-entrypoint.sh
